test(mongodb-mcp-remote): rename to mock remote, simplify comments, support SSE+JSON (MCP-538)
- Rename the test double from 'fake' to 'mock' (file, type, function, literals).
- Rewrite comments in plain language; add a reason to the eslint-disable.
- Mock remote now defaults to SSE (matching cloud-dev per MCP-542) with a
  responseMode: 'json' option; notifications return 202.
