+++ packages/mongodb-mcp-remote/src/tokenManager.test.ts
@@ -1,25 +1,19 @@
 import { describe, it, beforeEach, afterEach, vi } from "vitest";
 
 /**
- * Unit tests for the wrapper's token manager.
- *
- * Unit tests test ONE piece in isolation with everything around it faked.
- * Here we fake the network entirely, so these tests are instant and
- * deterministic — no fake server, no subprocess.
  *
  * Default to mocking the `oauth4webapi` module, consistent with the MCP server's
  * existing auth unit test (tests/unit/common/auth/clientCredentials.test.ts).
+ *
  * TODO: if the wrapper's token manager ends up calling `fetch` directly instead of
  * using `oauth4webapi`, switch to mocking `fetch` (vi.stubGlobal("fetch", ...)).
  *
- * NOTE: every test here is `it.todo` until the token manager exists in
- * packages/mongodb-mcp-remote/src/. Once you add it (e.g. `tokenManager.ts`
- * exporting a `TokenManager` class), import it here and flesh these out.
- * (Add `oauth4webapi` as a package dependency when wiring this up.)
+ * TODO: every test here is `it.todo` until the token manager exists in
+ * packages/mongodb-mcp-remote/src/. Tests will be implemented in MCP-539.
  */
 
-// Example of how you'll fake the network with oauth4webapi (mirrors
-// clientCredentials.test.ts). Uncomment and adapt once the token manager exists.
+// Example of mocking the network with oauth4webapi (similar to
+// clientCredentials.test.ts).
 //
 // import * as oauth from "oauth4webapi";
 // import { TokenManager } from "./tokenManager.js";
